Split out AS2 Content-Type utils
